requirements.txt
kafka-python=
apache-flink=1.16.1